ArtGAN/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
enum34==1.1.6             # via imageio
funcsigs==1.0.2           # via mock
futures==3.3.0            # via imageio
imageio==2.5.0
mock==3.0.5               # via tensorflow-gpu
numpy==1.16.4
pillow==6.1.0             # via imageio
protobuf==3.9.0           # via tensorflow-gpu
scipy==1.2.2
six==1.12.0               # via mock, protobuf, tensorflow-gpu
tensorflow-gpu==1.1.0
werkzeug==0.15.5          # via tensorflow-gpu
wheel==0.33.4             # via tensorflow-gpu

# The following packages are considered to be unsafe in a requirements file:
# setuptools==41.0.1        # via protobuf
